Attempt better compatibility with non-TLS1.1 https servers.
Try to use the (not-yet-widely-deployed) SSL3_RECORD_VERSION option in
gnutls to make it handshake according to the backward-compatibility
recommendations in the TLS spec. If that fails, fall back to just
disabling TLS1.1. Should hopefully help us to work better with servers
that freak out if you even suggest using TLS1.1.

Fix based on a Pidgin patch from Martin von Gagern.

http://bugzilla.gnome.org/show_bug.cgi?id=581342