Don't emit "authenticate" until a message is actually sent and fails.
In some cases, we're reasonably sure it's going to fail if we don't
emit authenticate, but forcing callers to deal with both the
pre-sending and post-sending cases in their authenticate handlers
makes writing them tricky. (In particular, calling
soup_message_io_pause() in the pre-sending case won't work, since the
message hasn't actually started yet.) This also solves the problem
that we were previously emitting "authenticate" both before and after
the message was sent in some cases.

Add a regression test for this case (written by Gustavo Noronha, who
figured out the circumstances where the bug was happening), and tweak
the existing regression tests to match the new rules.

http://bugzilla.gnome.org/show_bug.cgi?id=583462