Implement content sniffing
The implementation is based on the draft spec on Content-Type
Processing Model (draft-abarth-mime-sniff-01). It is a spinoff from
the HTML5 spec.

Soup now provides a SoupContentSniffer session feature, which hooks
into the message I/O, and delays emissions of the got-chunk signal to
be able to figure out the Content-Type of messages from the actual
content received, in some cases.

GIO is also used to sniff content, whenever the spec allows further
sniffing.

http://bugzilla.gnome.org/show_bug.cgi?id=572589