SoupSession: Add http-aliases and https-aliases properties
Currently SoupSession treats all URI schemes except "https" as aliases
for "http", and some apps depend on this. (Eg, iTunes sometimes
returns redirects involving "daap" URIs, which Rhythmbox needs to
treat as "http".) Unfortunately, this also means that it mishandles
redirects to, eg, ftp.

The http-aliases and https-aliases properties allow an app to
explicitly indicate which URI schemes should be considered aliases for
http and https, with other schemes considered to be unknown and
unhandled.