Avoid infinite loop trying to connect to host when the network is down.
If a connection attempt to a host fails, fail all of the messages
queued for that host.

Previously we would leave them in the queue if there was already an
existing connection to the host (hoping that we'd be able to reuse
that connection for the other messages later), but this meant that if
there was a single connection to a host open, with a message running
on it, and at least one additional queued message to the same host,
and then the network went down, libsoup would get stuck in a loop
repeatedly trying and failing to open a second connection (with the
first connection never getting a chance to time out).

http://bugzilla.gnome.org/show_bug.cgi?id=592492