Be less strict when checking whether to apply "text or binary"
We currently follow the spec to the letter, and only apply the text or
binary algorithm when we find one of three specific content types,
parameters included. It turns out we also want to run that check with
almost any other text/plain variation, such as "text/plain;
charset=utf-8" (only the upper-cased variant is considered today). The
new behaviour matches chromium, and the spec is going to be updated,
according to Adam Barth.

https://bugzilla.gnome.org/show_bug.cgi?id=611502