SoupProxyResolverGNOME: improve behavior with more-recent libproxy
SoupProxyResolverGNOME had a nasty workaround for the fact that
libproxy 0.2 would crash inside GConf if called from a thread other
than the main one. Unfortunately, the workaround involves leaking
proxy settings into the environment, which is particularly bad since
gnome-panel uses libsoup (via libgweather via intlclock) and so the
environment can become polluted for everything started from the panel.

In libproxy 0.3 and later, the workaround is no longer necessary, so
remove it if libproxy is new enough.

https://bugzilla.gnome.org/show_bug.cgi?id=603285