SoupSocket: tweak GMainContext handling
GSocketClient may take multiple steps, in which case we need to make
sure that the socket's async_context is the thread-default context
throughout all of them.

This will break horribly if someone tries to send things off to
multiple different contexts from a single thread, but that was never
supposed to have worked, and always had race conditions anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=635101