Fix handling of connections that time out while IN_USE
To avoid looping, we only resend a request on
unexpected-connection-close if it was sent on a connection that has
previously been succesfully used. But we were only setting the
"successfully" used flag after a message *completed*, so if a message
just got restarted, and the connection got closed before the second
sending, then the connection was seen as having never been used, and
so the message wouldn't get re-sent.

Fix this by marking the connection as having been used if a message it
is sending gets restarted.

https://bugzilla.gnome.org/show_bug.cgi?id=660057