soup-cache: correctly handle conditional requests not returning 304
If we do not get a 304 after a conditional request we have to refresh the
cache contents for that resource. We should use the original SoupMessage to
retrieve the new data from the resource (we were incorrectly using the
SoupMessage used for the conditional request).

The current strategy is clearly suboptimal as we ask again for the resource
instead of directly using the data from the response of the conditional
request. This will change once the cache is integrated into the streams
stack.