SoupSocket: don't emit "readable" on disconnect for blocking sockets
"readable" and "writable" are only supposed to be emitted for
non-blocking sockets, but we were emitting "readable" on disconnect
for all sockets.

In libsoup 2.38 and earlier, this could cause a crash if a message in
a SoupSessionSync failed at certain points, since the code wasn't
expecting to end up in io_read(). The current code no longer has this
problem (since soup-message-io no longer uses the SoupSocket signals),
but other external users of SoupSocket might.