SoupCacheInputStream: new input stream filter that writes to the cache
The SoupCacheInputStream will be added to the stream stack as any other
GPollableInputStream. It will transparently read data from the underlying
stream and will pass it unmodified to the upper level.

Apart from that the stream writes everything it reads to a local file. Once
the caching finishes a callback will be called. Caching may be cancelled at
any point by providing a GCancellable.

https://bugzilla.gnome.org/show_bug.cgi?id=682112