SoupSession: make queue items hold a ref on the session
Historically, messages sent via soup_session_queue_message() did not
hold a ref on the session, and unreffing the session when messages
were in progress would cancel them. This is ugly internally though,
and at any rate, depends on being able to know for sure that no one
else is secretly holding a ref on the session.

So, fix it so that pending messages hold a ref on the session, and you
need to use soup_session_abort() before unreffing if you want to
cancel everything.

If it turns out that this breaks things it can be reverted before
2.40...