SoupContentSnifferStream: added can_poll() implementation
Done in terms of the base_stream can_poll() interface. It was not needed to
read resources from network as the underlying streams were all pollable.
This will be used later to sniff locally cached resources. The file streams are
not pollable so we must implement the can_poll() in order to use "normal"
read() calls instead of the read_nonblocking().

https://bugzilla.gnome.org/show_bug.cgi?id=682112