SoupCache: add SoupContentProcessor support
SoupCache now implements the SoupContentProcessor interface. This means that
soup-message-io will ask it to wrap a given base stream. By doing that the
cache will backup up every byte readed from that base stream into a local
file using a SoupCacheInputStream.

Apart from that, this removes all the old resource writing infrastructure
that was moved to the new SoupCacheInputStream. The SoupCache just needs
to pass a callback to the stream in order to be notified once the caching
operation finishes.

This also involves moving the caching decision from the "got-headers"
callback to the _wrap_input() implementation.

https://bugzilla.gnome.org/show_bug.cgi?id=682112