SoupConnection: hide a special case
Previously SoupConnection required SoupSession to request that the
connection's state be changed from IN_USE to IDLE after a CONNECT,
even though this would not actually result in the connection's state
changing, and even though SoupSession would have broken if the state
had changed. Fix this up so that SoupSession can just leave the
connection IN_USE in this case, and SoupConnection deals with fixing
up its internal state accordingly.