SoupProxyResolver: remove evil hack
Back when I deprecated SoupProxyResolver in favor of
SoupProxyURIResolver, I added an evil hack so that if anyone created a
SoupProxyResolver implementation, we'd automatically add a
SoupProxyURIResolver implementation on top of it (so that SoupSession
only needed to worry about SoupProxyURIResolver).

Anyway, (a) it's evil, (b) I'm pretty sure no one else ever
implemented a SoupProxyResolver anyway, and (c) if they did, they
really ought to have migrated to SoupProxyURIResolver by now.

So, remove the evil hack so that desrt can drop support for adding
interfaces after class_init.

https://bugzilla.gnome.org/show_bug.cgi?id=687659