soup-auth-manager: split out connection handling
Add a new SoupConnectionAuth class to help with connection tracking,
and make SoupAuthNTLM a subclass of it. Allow a single SoupAuthNTLM to
carry state information about multiple connections.

Make SoupSession store the SoupConnection a SoupMessage is associated
with on the message, and use that from SoupConnectionAuth rather than
tracking sockets by hand like SoupAuthManager had previously done.
Remove the connection tracking in SoupAuthManager, since it is no
longer needed.