soup-message: fix up add_header_handler()/add_status_handler() docs
soup_message_add_header_handler() and
soup_message_add_status_handler() used to have the semantics that the
handlers they added wouldn't be called if the message was cancelled by
a preceding signal handler. This feature accidentally got lost in
2.32, but no one appears to have noticed. Fix up the docs to
correspond to the current reality.

Also, fix part of the add_header_handler() docs that *never*
corresponded to reality.