Rename all remaining "SoupMessage *req"s to "msg"
Originally (ie, 12 years ago last week) the HTTP message type in soup
was called SoupRequest, and most SoupRequest variables were called
"req". Shortly after, SoupRequest was renamed to SoupMessage, but none
of the existing "req"s were renamed, and in fact, new ones kept
getting added. Eventually, "msg" became the standard name for a
SoupMessage variable, but a handful of "req"s have managed to survive
to this day. With the increasing integration of the modern-day
SoupRequest, this has gone from "inconsistent" to "confusing", so fix
all the remaining stray "SoupMessage *req"s.