Revert "SoupConnection: disconnect on failed connections/tunnels"
Disconnecting clears priv->sock, and in some cases, the caller may
want to look at that after the failed connection. So switch this back
to having the session do the disconnecting itself (which is simpler
now due to various SoupSession rewrites), and remove one inconsistent
case where the connection was disconnecting itself before.

This reverts commit 21d707d66460233596a04f6dc99c945d2dbc3085.