Use GProxyResolver for all proxy resolving
Finish up the GProxyResolver porting started in 2.42 by removing all
the non-GProxyResolver proxy code from SoupSocket and SoupConnection
and implementing SoupSession:proxy-uri in terms of
GSimpleProxyResolver. If the user adds a SoupProxyResolverDefault to
the session, just use its GProxyResolver directly instead. If the user
adds any other kind of SoupProxyURIResolver to the session, use the
new SoupProxyResolverWrapper, which is a GProxyResolver that
translates from SoupProxyURIResolver.

https://bugzilla.gnome.org/show_bug.cgi?id=680273