Fix build on Windows, deprecate soup-portability.h
soup-portability.h created portability problems (ha!) because the
windows headers do "#define interface struct", conflicting with
variable names elsewhere.

The only thing that needed soup-portability.h in the public headers
was soup-address.h, which uses struct sockaddr. But we can just do a
forward declaration of that type rather than actually pulling in the
headers.

In git master, we made soup-address.h stop including
soup-portability.h, but that might possibly break some other modules
that depended on implicitly getting network includes as a side effect
of including soup.h, and we shouldn't be doing that in the stable
branch. So instead, we just tweak things so that it doesn't get
included from within libsoup, but still does externally.

https://bugzilla.gnome.org/show_bug.cgi?id=692134