Bug 703186 - fallback from automatic NTLM auth not working
There were two issues. Firstly, when the automatic NTLM authentication
failed we were setting conn->state to SOUP_NTLM_FAILED. So although we
would emit the 'authenticate' signal and obtain a password, we'd never
actually *try* it. We have the SOUP_NTLM_SSO_FAILED state for this, which
makes it work.

Secondly, the resulting 'authenticate' signal we being emitted with the
'retrying' property set, even when it was the first time we actually asked
for a password. This was confusing Evolution (bug 703181 comments 1-3).

Fix the latter problem by making soup_auth_ntlm_is_connection_ready()
return TRUE when the state is SOUP_NTLM_SSO_FAILED. That makes
auth_got_headers() do the right thing, and Evolution is happy again.

With this fixed, Evolution will now automatically try to connect to an
Exchange server using automatic NTLM authentication, and will fall back
to asking for a password only if it actually needs to.

(cherry picked from commit 2b9345730cd95ea9efb4e1b78d40574f6c316f96)