[soup-logger] On response, always print headers
SoupLogger won't print any information upon
response until the "got-body" or
"got-informational" signals are received. If the
response message has been cancelled or requeued
first, we miss the headers information, even when
we would have been already able to print it.

Now, upon response, if the logging has not yet
happened when the "finished" signal is triggered,
we do perform the logging then.

https://bugzilla.gnome.org/show_bug.cgi?id=703200